Refactor the macro setting element to avoid TeX encoding dangers
The previous version was ugly and dangerous. By collecting each git log
parameter in its own popen, and using tex.?print(-2, _) mode, we can avoid
the danger of non-catcode 10 and 12 characters. This has only removed the
danger, not the ugly. @FIXME: map ascii characters to the correct TeX
encoding, eg \ needs to map to \ not `.